Re-home the liquidation recovery helpers
We also take this time to split out DeriveAddress, which is more general
